requirements.txt
whisperx==3.3.4
python-dotenv
pandas
torch 
torchvision 
torchaudio